Content and Container
Content is a base class for named things, that must belong to a Container.
It has two attributes: name and belongs_to, which are both "inmutable".
Also, it checks that name is a valid Python identifier.

Container stores Content, and is itself a Content (has name and belongs_to).
It checks that only Content belonging to itself may be added, and that
there are no name collisions with previously stored Content.

Implemented Parameter and Reactant as Content. Removed states from Reactant.
It could be reimplemented as a subclass of Container.