Compartment as Container
Reimplemented Compartment as a subclass of Container.

Reactants, Parameters and Compartments should be
created from add_* methods, which automatically set
belongs_to to the current Compartment.

Reactions are kept outside of the Container machinery,
as they don't have a name.
Reactions can only be added if all Reactants and Parameters
belong to a Container or subContainers.

Universe is implemented as a special subclass of Compartment,
that belongs to None.