Moved common logic to BaseReaction
Subclasses of both SingleReaction and CompoundReaction are created by defining
a method (rhs and yield_reactions, respectively), with the proper annotations.

Subclasses derive annotations from there, save the generic Reactant
and Parameter names, and get created as dataclasses.

On subclass instantiation, __post_init__ validates arguments, unpacks
InReactionReactants and saves st_numbers, which are necessary to
recreate the instance (in case of copy).