Added dataclass decorator to Content subclasses
Content subclasses were instantiated as dataclasses from __init_subclass__,
which removed the hassle of adding a decorator with frozen=True and eq=False everywhere.

On the other side, it was less explicit about what was happening, and,
more importantly, it wasn't supported by static-code analyzers,
which now show the __init__ signature.