Prepared classes for automatic naming and ownership
Removed dataclasses, since they don't allow for keyword-only parameters,
and we were using increasingly less of its features.

Renamed belongs_to to parent, which is assigned when Content is added to a Conteiner.
Content.__set_name__ will assign name and parent in the new class-based API.

Removed Universe, which was a Compartment whose parent was None,
which added more confusion to the API, and is no longer needed
since the default parent is now None until added to a Compartment.

Removed default values of 0 in Parameters and Species, to be more explicit.