Metaclass for class-based Models
Compartment now has metaclass Model, which treats differently Content,
that is, Species, Parameters and Compartments, in its class body definition.
They are collected separately, not added to __dict__, but to _contents.

Model metaclass checks for duplicate Content, or collisions with inherited
Content that has not been explicitly marked with override=True.

Inheritance rules for Content are different from usual Python inheritance:
all Content is copied from the classes which are direct ancestors.