Added binding operator for Species
Species.__and__ is used to create a new Species, with zero initial concetration.

If A and B are Species, A&B is a Species belonging to the common parent of A and B.
If there's no common parent, then an error is raised.
If there is, it is either retrieved, or created and added to the common parent.
Also, A&B == B&A.

Fixed Content._common_parent, which returned the Content itself if only
multiple references of a Content were passed as arguments.