New new Reactions
Several changes in reactions, some of slight changes in API names,
and big change in SingleReaction definitions.

The old BaseReaction and CompoundReaction are merged into Reaction,
which defines a reactions method that yields other reactions.
Reactions are recursively "unpacked" into SingleReactions with the
helper method single_reactions. SingleReaction is a subclass of Reaction
which yields itself, ending the recursive behavior.

SingleReaction has major changes. Breaks down the rhs method in two components:
reaction_rate and reaction_balance. Reaction rate is similar to the old rhs,
but returns a float, instead of a tuple, which is the rate at which the reaction
occurs. Reaction balance specifies which Species are consumed (reactants) and
which are created (products). The rhs will then be built from reaction_rate
and reaction_balance.