Compartments: add and override reactions in class-based API
Model._reactions changes from OrderedSet[BaseReaction] to
Dict[ReactionBalance, SingleReaction].

There are two different levels of equivalence for SingleReactions:
- Reactions are equivalent if they have the same ReactionBalance,
but possibly different parameters or reaction_rate formula.
- Reactions are equal if they are instances of the same reaction,
with the same ReactionBalance and parameters.

We want to have a unique reaction in a given Model for a given
ReactionBalance, hence, the current _reactions dict. The Model
raises an error if an equivalent reaction tries to be added.
But we want to be able to override a given reaction to change
its parameters, hence the override flag. But we don't want to
raise errors in multiple inheritance of Models when reactions
are equal.