Simulator: Builder and Output
There's now a unique Simulator class that takes a model and solver,
builder and output classes.

The Builder class is in charge of building the model's RHS from the
model's reactions. There are two Builder classes, numpy and numba
builders. Numba builder is slower to build the RHS, but is necessary
to use numbakit-ode solvers.

Output class is in charge of "formatting" the output of a simulation.
The default is "pandas", which returns a DataFrame with columns named
according to the corresponding species.