Changing class-based API definition of Parameter and Species
Instead of using directly an instance of Species/Parameter,
we use an annotation and a float.

There are three benefits:

1. It is closer to the method-based builder API,
where methods such as `add_species` receive a number, not a Species instance.

2. It resolves `test_add_external_species` in test_single_level.py for the Static case,
where we couldn't identify if X was a new instance inside the Model
or an already existing external instance.

3. It forces all Species and Parameters to be annotated,
without duplicating code (ie, not `A: Species = Species(0)`).
Then, all will appear in the calling signature which will
be used to override components.